Modified the DataBrowser to accept user entries for directories
